requirements.txt
django = *
djangorestframework = *
dj-rest-auth = *
django-allauth = *
django-filter = *
django-cors-headers = *
django-guardian = *
drf-yasg = *